m3space/m3space_test.go
package m3space

import (
	"testing"
	"github.com/stretchr/testify/assert"
)

func TestSpace(t *testing.T) {
	DEBUG = true

	assert.Equal(t, int64(9), SpaceObj.max)
	assert.Equal(t, 0, len(SpaceObj.nodes))
	assert.Equal(t, 0, len(SpaceObj.connections))
	assert.Equal(t, 0, len(SpaceObj.events))
	assert.Equal(t, 0, len(SpaceObj.Elements))

	SpaceObj.CreateStuff(3)

	assert.Equal(t, int64(3), SpaceObj.max)
	// Big nodes = (center + center face + middle edge + corner) * (main + 3)
	nbNodes := (1 + 6 + 12 + 8) * 4
	assert.Equal(t, nbNodes, len(SpaceObj.nodes))
	assert.Equal(t, nbNodes+35, len(SpaceObj.connections))
	assert.Equal(t, 4, len(SpaceObj.events))
	assert.Equal(t, 2*nbNodes+35+6, len(SpaceObj.Elements))
	assert.Equal(t, TickTime(0), SpaceObj.currentTime)
	assertOutgrowth(t, 4)
	assertOutgrowthDistance(t,map[EventID]int{0:1,1:1,2:1,3:1})

	SpaceObj.ForwardTime()
	// Same elements just color changes
	assert.Equal(t, nbNodes, len(SpaceObj.nodes))
	assert.Equal(t, nbNodes+35, len(SpaceObj.connections))
	assert.Equal(t, 4, len(SpaceObj.events))
	assert.Equal(t, 2*nbNodes+35+6, len(SpaceObj.Elements))
	assert.Equal(t, TickTime(1), SpaceObj.currentTime)
	assertOutgrowth(t, 4+(4*3))
	assertOutgrowthDistance(t,map[EventID]int{0:3,1:3,2:3,3:3})

	SpaceObj.ForwardTime()
	// Same elements just color changes
	assert.Equal(t, nbNodes, len(SpaceObj.nodes))
	assert.Equal(t, nbNodes+35, len(SpaceObj.connections))
	assert.Equal(t, 4, len(SpaceObj.events))
	assert.Equal(t, 2*nbNodes+35+6, len(SpaceObj.Elements))
	assert.Equal(t, TickTime(2), SpaceObj.currentTime)
	assertOutgrowth(t, 4+(4*3)+(4*3)+2)
	assertOutgrowthDistance(t,map[EventID]int{0:3,1:3,2:3,3:5})

	SpaceObj.ForwardTime()
	// Same elements just color changes
	assert.Equal(t, nbNodes, len(SpaceObj.nodes))
	assert.Equal(t, nbNodes+35, len(SpaceObj.connections))
	assert.Equal(t, 4, len(SpaceObj.events))
	assert.Equal(t, 2*nbNodes+35+6, len(SpaceObj.Elements))
	assert.Equal(t, TickTime(3), SpaceObj.currentTime)
	assertOutgrowth(t, 4+(4*3)*4+1)
	assertOutgrowthDistance(t,map[EventID]int{0:4,1:6,2:4,3:9})
}

func assertOutgrowth(t *testing.T, expect int) {
	nbOutgrowth := 0
	for _, evt := range SpaceObj.events {
		nbOutgrowth += len(evt.outgrowths)
	}
	assert.Equal(t, expect, nbOutgrowth)
	nbOutgrowth = 0
	for _, node := range SpaceObj.nodes {
		nbOutgrowth += len(node.E)
	}
	assert.Equal(t, expect, nbOutgrowth)
}

func assertOutgrowthDistance(t *testing.T, topOnes map[EventID]int) {
	for _, evt := range SpaceObj.events {
		nbTopOnes := 0
		for _, eo := range evt.outgrowths {
			if eo.distance == Distance(SpaceObj.currentTime-evt.created) {
				assert.Equal(t, eo.state, EventOutgrowthLatest, "Event outgrowth state test failed for evtID=%d node=%v . Should be latest", evt.id,*(eo.node))
				nbTopOnes++
			} else {
				assert.Equal(t, eo.state, EventOutgrowthOld, "Event outgrowth state test failed for evtID=%d node=%v . Should be old", evt.id,*(eo.node))
			}
		}
		assert.Equal(t, topOnes[evt.id], nbTopOnes, "NB top ones expected failed for evtID=%d", evt.id)
	}
}
